fix(vbc/interpreter): zero-extend UInt32 typed-array reads
Typed-array reads through both `handle_get_index` (GetE opcode, TypeId::U32
path) and `DerefRaw` (used by the NewTypedArray write-then-read pipeline
for `[UInt32; N]` locals) cast the raw slot to `*const i32` and then
`as i64`, which sign-extends any element whose bit 31 was set. The
resulting value carried `0xFFFFFFFF` in its upper 32 bits through every
subsequent shift/XOR, corrupting the computation.

The canonical symptom was CRC32 diverging from zlib: a per-byte table
whose entries are genuinely unsigned 32-bit values (e.g.
`TABLE[0x7A] = 0xB0D09822`) came back as `0xFFFFFFFF_B0D09822` when
indexed, and `>> 8` on that value dragged `0xFF` down into the low byte
of the next state word.

Fix: read via `*const u32` (and the 2-byte path via `*const u16`)
so the NaN-box slot holds the zero-extended bit pattern. Callers that
genuinely want signed semantics can truncate with `as i32` at the use
site — zero-extension is the invariant-preserving default for raw
unsigned storage and matches the policy in the FatRef branch.

Regression test: vcs/specs/L1-core/types/unsigned_shr_logical.vr.
The matching `>>` unsigned-emits-Ushr codegen change landed earlier
in 71cce67e (session handoff).

+++ crates/verum_vbc/src/interpreter/dispatch_table/handlers/ffi_extended.rs
@@ -564,12 +564,25 @@ pub(in super::super) fn handle_ffi_extended(state: &mut InterpreterState) -> Int
             // valid for reads of `size` bytes — this mirrors Rust's raw-pointer
             // semantics and matches the AOT lowering. `read_unaligned` handles
             // arbitrary alignment.
+            //
+            // Extension policy (root fix for Issue #2 — continuation from
+            // `handle_get_index`): `DerefRaw` is emitted by typed-array
+            // reads and by any intrinsic that lowers to a raw memory read.
+            // Widths 1/2/4 are *zero-extended* into the i64 NaN-box slot so
+            // that a `[UInt32; N]` (or `[UInt8; N]` / `[UInt16; N]`) element
+            // whose high bit is set is preserved as an unsigned value.
+            // Sign-extension via `as i32 as i64` corrupted the upper 32 bits
+            // of every u32 with bit 31 set and was the root cause of the
+            // CRC32 divergence against zlib. Callers that actually need
+            // signed semantics for a raw read can truncate (`as i32` etc.)
+            // at the use site — zero-extension is the invariant-preserving
+            // default for unsigned raw I/O, which is the vastly common case.
             let value = unsafe {
                 match size {
-                    1 => *ptr as i8 as i64,
-                    2 => std::ptr::read_unaligned(ptr as *const i16) as i64,
-                    4 => std::ptr::read_unaligned(ptr as *const i32) as i64,
-                    8 => std::ptr::read_unaligned(ptr as *const i64),
+                    1 => *ptr as i64,                                                 // u8 → i64 (zero-extend)
+                    2 => std::ptr::read_unaligned(ptr as *const u16) as i64,          // u16 → i64
+                    4 => std::ptr::read_unaligned(ptr as *const u32) as i64,          // u32 → i64
+                    8 => std::ptr::read_unaligned(ptr as *const i64),                 // 8 bytes fill the slot
                     _ => return Err(InterpreterError::InvalidOperand {
                         message: format!("invalid deref size: {}", size),
                     }),
